+++ app/pages/analyst/application/[applicationId]/assessments/gis.tsx
@@ -55,7 +55,11 @@ const GisAssessment = ({
   const showApplicationGisData = useFeature('show_application_gis_data').value;
 
   return (
-    <Layout session={session} title="Connecting Communities BC">
+    <Layout
+      session={session}
+      title="Connecting Communities BC"
+      provisionRightNav
+    >
       <AnalystLayout query={query}>
         <AssessmentsTabs />
         {showApplicationGisData && <ApplicationGisData query={query} />}