+++ app/pages/analyst/application/[applicationId]/assessments/screening.tsx
@@ -56,7 +56,11 @@ const ScreeningAssessment = ({
   const { applicationByRowId, session } = query;
 
   return (
-    <Layout session={session} title="Connecting Communities BC">
+    <Layout
+      session={session}
+      title="Connecting Communities BC"
+      provisionRightNav
+    >
       <AnalystLayout query={query}>
         <AssessmentsTabs />
         <StyledInfoBarDiv>